+++ src/Mixin/LocalizedRoute.php
@@ -51,7 +51,6 @@ class LocalizedRoute
     {
         return function (string $locale): ?Route {
             /** @var Localized&Route $this */
-
             if (! in_array($locale, $this->lang(), true)) {
                 return null;
             }